refactor and add hover functions
