+++ extension/src/tree-sitter/index.js
@@ -1,6 +1,6 @@
 const Parser = require("web-tree-sitter");
 const path = require("path");
-const HoverFuncs = require("./hover.js");
+const getHoverPopover = require("./hover/index.js");
 
 /** @typedef {import('./types.d.ts').Idents} Idents */
 
@@ -26,52 +26,14 @@ const getParser = () => parser;
 /**
  * TODO: add for idents:
  *  - br, br_if
- *  - call_indirect (table)
- *  - type
  *  - allow index idents
  */
-
-const queries = /** @type {const} */ ({
-	type: `(module_field_type
-      (identifier) @ident 
-      (type_field 
-        (func_type (func_type_params) @params)*
-        (func_type (func_type_results) @result)*
-      )
-    )`,
-
-	table: `(module_field_table identifier: (identifier) @ident 
-      (table_fields_type (table_type
-        (limits) @limits
-        (ref_type) @ref_type
-      )) 
-    )`,
-});
 
 /** @param {Parser.SyntaxNode} node */
 function getHoverData(node) {
 	if (!languageWasm) throw Error("not initialized yet!");
 
-	if (node.type !== 'identifier' || !node.parent) return;
-
-	const parent = node.parent;
-
-	if (parent.type === "module_field_func") {
-		return HoverFuncs.getFunctionHoverString(languageWasm, parent);
-	}
-
-	if (parent?.parent?.type === "instr_plain") {
-		const surroundingText = parent.parent.text;
-		if (surroundingText.startsWith('global.')) {
-			return HoverFuncs.getGlobalHoverString(languageWasm, node);
-		}
-		if (surroundingText.startsWith('local.')) {
-			return HoverFuncs.getLocalHoverString(languageWasm, node);
-		}
-		if (surroundingText.startsWith('call')) {
-			return HoverFuncs.getCallHoverString(languageWasm, node);
-		}
-	}
+	return getHoverPopover(languageWasm, node);
 }
 
 module.exports = {
